refactor: split folder-graph into focused sub-module
The 328-line folder-graph.ts was doing six distinct things: Tarjan SCC,
folder-cycle + root-internal-cycle diagnostics, mesh density, distant
imports (already extracted), cross-domain sibling imports, and
cross-layer imports. Split into a folder-graph/ sub-module:

  imports/folder-graph/
    index.ts          50 lines  — orchestrator (checkFolderGraph)
    cycles.ts        143 lines  — Tarjan SCC + cycle + root-internal-cycle
    mesh.ts           82 lines  — production projection + density + mesh
    cross-folder.ts  136 lines  — cross-domain-sibling + cross-layer

production-folder-graph.ts is gone — its content lives in mesh.ts now,
right next to the package-mesh check that's the only thing that uses
the production-only projection. No more architecture-exception
directive needed.

Updated consumers: imports/index.ts re-exports from folder-graph/index.js;
test-support/helper-fixtures.ts likewise.
